+++ qf_lib/backtesting/execution_handler/commission_models/ib_commission_model.py
@@ -23,7 +23,7 @@ class IBCommissionModel(CommissionModel):
     https://www.interactivebrokers.co.uk/en/index.php?f=1590&p=stocks1
     """
 
-    def calculate_commission(self, fill_quantity: int, fill_price: float) -> float:
+    def calculate_commission(self, fill_quantity: float, fill_price: float) -> float:
         fill_quantity = abs(fill_quantity)
         commission = max(1.0, min(0.005 * fill_quantity, 0.01 * fill_price * fill_quantity))
